wip: have IRQ invocations depend on whether the kernel is configured to support setting the trigger.
There exists a configuration option `HAVE_SET_TRIGGER` that needs to
be set in order to have an invocation such as `seL4_IRQControl_GetTrigger`
work. This configuration option is currently defined in the IRQ
controller source code, and is not exposed for any code outside the
kernel to use. Because of this, this patch does not currently work.